Move State class from Kit to its own file
